task: Cleanup test script by splitting out test elements
