+++ test/lit-mobx.test.ts
@@ -10,109 +10,18 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { MobxLitElement } from '../';
-import { observable, computed } from 'mobx';
-import {
-    html,
-    customElement,
-    property,
-    TemplateResult,
-    PropertyValues,
-} from 'lit-element';
+import { html } from 'lit-element';
 import { expect, fixture, elementUpdated } from '@open-wc/testing';
-import { nothing } from 'lit-html';
+import {
+    TestMobxLitElementUpdatesArrays,
+    TestMobxLitElementUpdateChange,
+    TestMobxLitElementUpdatedChange,
+    TestMobxLitElementFirstUpdatedChange,
+} from './test-elements';
 
-@customElement('test-mobx-lit-element-updates-arrays')
-export class TestMobxLitElementUpdatesArrays extends MobxLitElement {
-    @property({ attribute: false })
-    @observable
-    public observableArray?: string[];
+// import elements for side effects
+import './test-elements';
 
-    @property()
-    public normalProperty = 'My Array';
-
-    @computed
-    private get arrayLength(): string {
-        return `Array length = ${
-            this.observableArray ? this.observableArray.length : 'N/A'
-        }`;
-    }
-
-    public render(): TemplateResult {
-        return html`
-            <p>${this.normalProperty}</p>
-            <p>
-                ${this.observableArray
-                    ? this.observableArray.join(', ')
-                    : nothing}
-            </p>
-            <p>${this.arrayLength}</p>
-        `;
-    }
-}
-
-@customElement('test-mobx-lit-element-update-change')
-export class TestMobxLitElementUpdateChange extends MobxLitElement {
-    @observable
-    public observableValue: number = 1;
-
-    @observable
-    private derivedValue: number = 1;
-
-    public update(changedProperties: PropertyValues) {
-        super.update(changedProperties);
-        this.derivedValue = this.observableValue * 2;
-    }
-
-    public render(): TemplateResult {
-        return html`
-            <p>${this.observableValue}</p>
-            <p>${this.derivedValue}</p>
-        `;
-    }
-}
-
-@customElement('test-mobx-lit-element-updated-change')
-export class TestMobxLitElementUpdatedChange extends MobxLitElement {
-    @observable
-    public observableValue: number = 1;
-
-    @observable
-    private derivedValue: number = 1;
-
-    public updated(changedProperties: PropertyValues) {
-        super.updated(changedProperties);
-        this.derivedValue = this.observableValue * 3;
-    }
-
-    public render(): TemplateResult {
-        return html`
-            <p>${this.observableValue}</p>
-            <p>${this.derivedValue}</p>
-        `;
-    }
-}
-
-@customElement('test-mobx-lit-element-first-updated-change')
-export class TestMobxLitElementFirstUpdatedChange extends MobxLitElement {
-    @observable
-    public observableValue: number = 1;
-
-    @observable
-    private derivedValue: number = 1;
-
-    public firstUpdated(changedProperties: PropertyValues) {
-        super.firstUpdated(changedProperties);
-        this.derivedValue = this.observableValue * 4;
-    }
-
-    public render(): TemplateResult {
-        return html`
-            <p>${this.observableValue}</p>
-            <p>${this.derivedValue}</p>
-        `;
-    }
-}
 describe('MobxLitElement', () => {
     it('updates arrays', async () => {
         const el = await fixture<TestMobxLitElementUpdatesArrays>(html`
